+++ authpass/test/cloud_storage/webdav_test.dart
@@ -27,12 +27,12 @@ void main() {
   _logger.fine('Starting webdav test for URL: $webDavUrl');
 
   group('simple webdav test', () {
-    WebDavProvider? provider;
+    late WebDavProvider provider;
     setUp(() async {
       // final env = await TestUtil.createEnv();
       provider = WebDavProvider(helper: CloudStorageHelperMock());
       // authenticate
-      final result = await provider!.startAuth((prompt) async {
+      final result = await provider.startAuth((prompt) async {
         if (prompt is UserAuthenticationPrompt<UrlUsernamePasswordResult,
             UrlUsernamePasswordPromptData>) {
           prompt.result(UrlUsernamePasswordResult(